macOS: extract SurfaceWrapper into its own file
SurfaceView.swift held the top-level SwiftUI view for a surface plus every
overlay it composes. Move SurfaceWrapper, and the SurfaceRepresentable it
wraps, into SurfaceWrapper.swift so the entry point is findable without
scrolling past ten overlay views.

SurfaceRepresentable is now `private`: SurfaceWrapper is its only user, so
file scope keeps it out of the way of the rest of the module.

Pure code motion otherwise.
